Allow quote_char to be specified from evaluator as template parameter, defaulted to double_quote_char. These changes make booleval a header only library.
