+++ include/booleval/token/tokenizer.hpp
@@ -33,6 +33,8 @@
 #include <vector>
 #include <string_view>
 #include <booleval/token/token.hpp>
+#include <booleval/utils/string_utils.hpp>
+#include <booleval/utils/split_range.hpp>
 
 namespace booleval {
 
@@ -44,12 +46,16 @@ namespace token {
  * Represents the mechanism for tokenizing the expressions, i.e. transforming
  * the expressions from a form of a string to the collection of tokens.
  */
+template <char quote_char = utils::double_quote_char>
 class tokenizer {
 public:
     tokenizer() = default;
     tokenizer(tokenizer&& rhs) = default;
     tokenizer(tokenizer const& rhs) = default;
-    tokenizer(std::string_view expression) noexcept;
+
+    tokenizer(std::string_view expression) noexcept
+        : expression_(expression) {
+    }
 
     tokenizer& operator=(tokenizer&& rhs) = default;
     tokenizer& operator=(tokenizer const& rhs) = default;
@@ -61,40 +67,52 @@ public:
      *
      * @param expression Expression to be tokenized
      */
-    void expression(std::string_view expression) noexcept;
+    void expression(std::string_view expression) noexcept {
+        expression_ = expression;
+    }
 
     /**
      * Gets the expression that needs to be tokenized.
      *
      * @return Expression to be tokenized
      */
-    [[nodiscard]] std::string_view expression() const noexcept;
+    [[nodiscard]] std::string_view expression() const noexcept {
+        return expression_;
+    }
 
     /**
      * Checks whether more tokens exist or not.
      *
      * @return True if there is more tokens, otherwise false
      */
-    [[nodiscard]] bool has_tokens() const noexcept;
+    [[nodiscard]] bool has_tokens() const noexcept {
+        return current_token_index_ < tokens_.size();
+    }
 
     /**
      * Passes the token by incrementing the current token index.
      */
-    void pass_token() noexcept;
+    void pass_token() noexcept {
+        current_token_index_++;
+    }
 
     /**
      * Gets the next token and increments the current token index.
      *
      * @return Next token
      */
-    [[nodiscard]] token const& next_token();
+    [[nodiscard]] token const& next_token() {
+        return tokens_.at(current_token_index_++);
+    }
 
     /**
      * Gets the next token without incrementing the current token index.
      *
      * @return Next token
      */
-    [[nodiscard]] token const& weak_next_token();
+    [[nodiscard]] token const& weak_next_token() {
+        return tokens_.at(current_token_index_);
+    }
 
     /**
      * Tokenizes the expression and transforms it into the collection of tokens.
@@ -104,13 +122,42 @@ public:
     /**
      * Clears the collection of tokens and sets the current index to zero.
      */
-    void reset() noexcept;
+    void reset() noexcept {
+        current_token_index_ = 0;
+    }
 
 private:
     std::string_view expression_;
     std::size_t current_token_index_{ 0 };
     std::vector<token> tokens_;
 };
+
+template <char quote_char>
+void tokenizer<quote_char>::tokenize() {
+    tokens_.clear();
+    reset();
+
+    constexpr auto parenthesis_symbols = parenthesis_symbol_expressions();
+    constexpr auto options =
+        utils::split_options::include_delimiters |
+        utils::split_options::split_by_whitespace |
+        utils::split_options::allow_quoted_strings;
+
+    auto delims = utils::join(std::begin(parenthesis_symbols), std::end(parenthesis_symbols));
+    auto tokens_range = utils::split_range<options, quote_char>(expression_, delims);
+
+    for (auto const& [quoted, index, value] : tokens_range) {
+        auto type = quoted ? token_type::field : map_to_token_type(value);
+
+        if (token_type::field == type) {
+            if (!tokens_.empty() && tokens_.back().is(token_type::field)) {
+                tokens_.emplace_back(token_type::eq, map_to_token_value(token_type::eq));
+            }
+        }
+
+        tokens_.emplace_back(type, value);
+    }
+}
 
 } // token
