+++ include/booleval/tree/expression_tree.hpp
@@ -45,7 +45,9 @@ namespace tree {
  * Represents a class for building an expression tree by using a recursive
  * descent parser method.
  */
-class expression_tree {
+template <char quote_char = utils::double_quote_char>
+class expression_tree
+{
 public:
     expression_tree() = default;
     expression_tree(expression_tree&& rhs) = default;
@@ -61,7 +63,9 @@ public:
      *
      * @return Root tree node
      */
-    [[nodiscard]] std::shared_ptr<tree::tree_node> root() noexcept;
+    [[nodiscard]] std::shared_ptr<tree::tree_node> root() noexcept {
+        return root_;
+    }
 
     /**
      * Builds the expression tree.
@@ -109,9 +113,135 @@ private:
     [[nodiscard]] std::shared_ptr<tree::tree_node> parse_terminal();
 
 private:
-    token::tokenizer tokenizer_;
+    token::tokenizer<quote_char> tokenizer_;
     std::shared_ptr<tree::tree_node> root_;
 };
+
+template <char quote_char>
+bool expression_tree<quote_char>::build(std::string_view expression) {
+    tokenizer_.reset();
+    tokenizer_.expression(expression);
+    tokenizer_.tokenize();
+
+    root_ = parse_expression();
+    if (nullptr == root_) {
+        return false;
+    } else if (tokenizer_.has_tokens()) {
+        root_ = nullptr;
+        return false;
+    }
+
+    return true;
+}
+
+template <char quote_char>
+std::shared_ptr<tree::tree_node> expression_tree<quote_char>::parse_expression() {
+    auto left = parse_and_operation();
+
+    auto const is_relational_operator =
+        tokenizer_.has_tokens() &&
+        tokenizer_.weak_next_token().is_one_of(
+            token::token_type::eq,
+            token::token_type::neq,
+            token::token_type::gt,
+            token::token_type::lt,
+            token::token_type::geq,
+            token::token_type::leq
+        );
+
+    if (is_relational_operator) {
+        return nullptr;
+    }
+
+    if (tokenizer_.has_tokens() && tokenizer_.weak_next_token().is_not(token::token_type::logical_or)) {
+        return left;
+    }
+
+    while (tokenizer_.has_tokens() && tokenizer_.weak_next_token().is(token::token_type::logical_or)) {
+        tokenizer_.pass_token();
+        auto logical_or = std::make_shared<tree::tree_node>(token::token_type::logical_or);
+
+        auto right = parse_and_operation();
+        if (nullptr == right) {
+            return nullptr;
+        }
+
+        logical_or->left = left;
+        logical_or->right = right;
+        left = logical_or;
+    }
+
+    return left;
+}
+
+template <char quote_char>
+std::shared_ptr<tree::tree_node> expression_tree<quote_char>::parse_and_operation() {
+    auto left = parse_parentheses();
+    if (nullptr == left) {
+        left = parse_relational_operation();
+    }
+
+    while (tokenizer_.has_tokens() && tokenizer_.weak_next_token().is(token::token_type::logical_and)) {
+        tokenizer_.pass_token();
+
+        auto logical_and = std::make_shared<tree::tree_node>(token::token_type::logical_and);
+
+        auto right = parse_parentheses();
+        if (nullptr == right) {
+            right = parse_relational_operation();
+        }
+
+        if (nullptr == right) {
+            return nullptr;
+        }
+
+        logical_and->left = left;
+        logical_and->right = right;
+        left = logical_and;
+    }
+
+    return left;
+}
+
+template <char quote_char>
+std::shared_ptr<tree::tree_node> expression_tree<quote_char>::parse_parentheses() {
+    if (tokenizer_.has_tokens() && tokenizer_.weak_next_token().is(token::token_type::lp)) {
+        tokenizer_.pass_token();
+        auto expression = parse_expression();
+        if (tokenizer_.has_tokens() && tokenizer_.weak_next_token().is(token::token_type::rp)) {
+            tokenizer_.pass_token();
+            return expression;
+        }
+    }
+
+    return nullptr;
+}
+
+template <char quote_char>
+std::shared_ptr<tree::tree_node> expression_tree<quote_char>::parse_relational_operation() {
+    auto left = parse_terminal();
+    if (tokenizer_.has_tokens()) {
+        auto operation = std::make_shared<tree::tree_node>(tokenizer_.next_token());
+        auto right = parse_terminal();
+        operation->left = left;
+        operation->right = right;
+        return operation;
+    }
+
+    return nullptr;
+}
+
+template <char quote_char>
+std::shared_ptr<tree::tree_node> expression_tree<quote_char>::parse_terminal() {
+    if (tokenizer_.has_tokens()) {
+        auto token = tokenizer_.next_token();
+        if (token.is(token::token_type::field)) {
+            return std::make_shared<tree::tree_node>(token);
+        }
+    }
+
+    return nullptr;
+}
 
 } // tree
